Defer reply if not already deferred
This has been tested with both deferred and non-deferred replies. Since there's no minimum limit on replying after deferral, it might be best to defer all replies in order to save on lines where methods would be changed depending on whether it's been deferred or not, especially since it ends up deferred later on down the line.